Implement capability free mode of operation
This commit implements oniux in a capability free fashion by utilizing
the `user_namespaces(7)` feature to gain capabilities on further
namespaces it "owns".  With this, we can performed a privileged
operation, such as the creation of TUN interfaces, without owning any
capabilities.

Many thanks to 7ppKb5bW for teaching on how this can be done, as well as
for providing an initial prototype with this feature enabled.