git: Ignore `/.cargo` all together
The file there was in fact only used for my own local development setup,
which involved cross-compiling from Darwin to Linux.  Personally, I
still have a file there but it does not need to be in the repository at
all at the moment.  If we make up our mind in the future, we can still
change that.